Modify cdc_dual_ports example to send large chunks of data to reproduce a rp4020 issue
This sends chunks of 532 bytes trash data to the tx fifo as long as there is room and the DTR bit is set.
Use the DTR to give you time to get a terminal open before the flood begins.
RP4020 will send data to the host for a random amount of time then "USBD Xfer Complete on EP 82 with 532 bytes" stop and the buffer fills up.
Will hang this way forever, host can no longer communicate with device.